third_party/WebKit/LayoutTests/cssom/cssimportrule-nested.html

<!DOCTYPE html>
<html>
  <head>
    <title>@import rule inside other rules should be ignored.</title>
    <meta name="flags" content="dom"/>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <style id="sheet1">
      @media all {
        @import "resources/import.css";
      }
    </style>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script>

      var doc = window.document;

      var sheet1 = doc.styleSheets[0];

      test(function(){
        var mediaRule = sheet1.cssRules[0];
        assert_equals(mediaRule.cssRules.length, 0);

      }, 'doc.styleSheets[0].cssRules[0].cssRules.length == 0');
    </script>
  </body>
</html>
